Added functionality to prevent a highlight from trying to delete itself if it has already been deleted. Now when a highlight is deleted is sets its highlight.deleted value to True.

+++ book.py
@@ -209,11 +209,12 @@ class Book:
     def deleteCompleteHighlights(self):
         
         for highlight in self.highlightList:
-            if highlight.truncated == False:
-                print("Deleting highlight:")
-                print(highlight.text + "\n")
-                highlight.delete()
-                time.sleep(2)
+            if highlight.deleted == False:
+                if highlight.truncated == False:
+                    print("Deleting highlight:")
+                    print(highlight.text + "\n")
+                    highlight.delete()
+                    time.sleep(2)
 
     def updateHighlights(self):
